Add files via upload
Currently just skips the spikes for the neruon that we are leaving. But essentially it is supposed to skip the cell entirely. so that means delete the place field and make the spikes from the one to be skipped to NaN and the next one to be one value lower